User Authentication: Request Authorization (#39)
Resolves #14 

## Changes
- Added authorization token to API requests made by fetchRequest() in
the frontend
- Modified firebase-admin set up in backend/src/firebase-admin.ts
- Implemented **validateUser()** in
backend/src/middleware/validation.ts. This is the middleware that checks
to see if a user making the API request is authenticated.
- Implemented **validateHousingLocator()** in
backend/src/middleware/validation.ts. This is the middleware that checks
to see if a user making the API request is a housing locator.

## Testing
- validateUser(): I logged in with my Google account and was able to
reach the end of the validator. At the end of the validator, I logged
the user info to console.
- validateHousingLocator(): I added this to the validateWith variable at
the end of validation.ts and got a frontend response when trying to log
in with my Google account that I wasn't a Housing Locator, I would get
an error response in the frontend. When I manually change the
isHousingLocator field through compass, I would not get the same error
as before.

## Confirmation of Change
- See "Testing" to see see how I tested the changes.

---------

Co-authored-by: Azhou2023 <anz008@ucsd.edu>
Co-authored-by: Philip Zhang <code@philipzha.ng>
Co-authored-by: AndrewZhou2023 <57163201+Azhou2023@users.noreply.github.com>