Navigation Bar + Login Redirect (#40)
## Tracking Info

Resolves #32  

## Changes

<!-- What changes did you make? -->

Added navigation bar for all pages and a rudimentary profile page.
Changed routes so that if a user isn't logged in, they are immediately
redirected to the login page no matter what path they try to visit.

## Testing

<!-- How did you confirm your changes worked? -->

Made sure that if there was no current user, all manual paths (valid or
invalid) would redirect to the login page. Also ensured that if a user
was logged in, they could navigate to the login page and back with no
barriers. For the nav bar, I tested that the paths were correct and that
the log out popup stopped any interaction with the rest of the screen
(and that logging out actually works and immediately redirects to the
login page).

## Confirmation of Change

<!-- Upload a screenshot, if possible. Otherwise, please provide
instructions on how to see the change. -->


![Screenshot_18](https://github.com/TritonSE/USHS-Housing-Portal/assets/57163201/9532dd42-a49d-45ef-b530-c292cc0a4933)

![Screenshot_19](https://github.com/TritonSE/USHS-Housing-Portal/assets/57163201/924a929b-e19e-442c-a7c0-39b502975900)

![Screenshot_20](https://github.com/TritonSE/USHS-Housing-Portal/assets/57163201/208140ae-79f9-417c-a2a8-2306b2446fde)

---------

Co-authored-by: Yashwanth Ravipati <yashravipatischool@gmail.com>