Simple unit browsing / Case Manager view (#43)
Resolves the frontend part of #20

- Implemented Filter components
- Implemented Unit Card component to display the unit information

Since the we couldn't work on the backend, there wasn't much testing to
be done. We mostly visually confirmed that things were looking the way
they should.

![image](https://github.com/TritonSE/USHS-Housing-Portal/assets/69605985/41e60ab9-1187-4474-9a93-a8e58af31f29)

![image](https://github.com/TritonSE/USHS-Housing-Portal/assets/69605985/0b64cef3-5e7d-4beb-ae9b-813e939a795d)

Every time a filter is applied, the query string that will be sent to
the backend will be printed in the console for testing purposes.

---------

Co-authored-by: nbpillai <nipillai@ucsd.edu>
Co-authored-by: Philip Zhang <code@philipzha.ng>